Readded .env.example file; organized components in folders; created BaseLayout component
