+++ .env.example
@@ -1,0 +1,3 @@
+ELASTIC_SEARCH_CLOUD_ID=""
+ELASTIC_SEARCH_API_KEY=""
+ELASTIC_SEARCH_INDEX="kibana_sample_data_logs"
